// src/lib/api/generated/model/v1ListPullRequestsResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * aisphere/options/v1/authz.proto
 * OpenAPI spec version: version not set
 */
import type { V1PullRequest } from './v1PullRequest';

export interface V1ListPullRequestsResponse {
  pullRequests?: V1PullRequest[];
  nextPageToken?: string;
}