// src/lib/api/generated/model/v1PullRequest.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Aisphere Hub API
 * OpenAPI spec version: 1.0.0
 */

export interface V1PullRequest {
  authorId?: string;
  createTime?: string;
  description?: string;
  id?: string;
  mergedSha?: string;
  mergedTime?: string;
  skillName?: string;
  sourceRef?: string;
  sourceSha?: string;
  state?: string;
  targetRef?: string;
  targetSha?: string;
  title?: string;
  updateTime?: string;
}